requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Web application
streamlit>=1.28.0

# AWS integration
boto3>=1.26.0
botocore>=1.29.0

# AI/ML models
ollama>=0.1.0

# Utilities
urllib3>=1.26.0
python-dotenv>=1.0.0